+++ kitty/simple_cli_definitions.py
@@ -644,7 +644,9 @@ default={layer}
 On a Wayland compositor that supports the wlr layer shell protocol, specifies the layer
 on which the panel should be drawn. This parameter is ignored and set to
 :code:`background` if :option:`--edge` is set to :code:`background`. On macOS, maps
-these to appropriate NSWindow *levels*.
+these to appropriate NSWindow *levels*. Note that on macOS, you can use :opt:`macos_ns_window_layer`
+for more fine control over the layer, using the :code:`--override` flag, for example:
+:code:`--override macos_ns_window_layer=NSPopUpMenuWindowLevel`.
 
 
 --config -c